+++ packages/braintree-utils/src/create-braintree-sdk.spec.ts
@@ -1,0 +1,8 @@
+import BraintreeSdk from './braintree-sdk';
+import createBraintreeSdk from './create-braintree-sdk';
+
+describe('createBraintreeSdk', () => {
+    it('instantiates braintree sdk', () => {
+        expect(createBraintreeSdk()).toBeInstanceOf(BraintreeSdk);
+    });
+});